requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg[binary]>=3.1
PyJWT>=2.8.0
